requirements.txt
falcon==2.0.0
gunicorn==20.0.4
requests==2.25.1
pytest==6.2.2
coverage==5.5
python-dateutil==2.8.1